move llm to tools
