+++ src/main/kotlin/org/jetbrains/research/testgenie/actions/GenerateTestsForClassLLM.kt
@@ -2,9 +2,9 @@ package org.jetbrains.research.testgenie.actions
 
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import org.jetbrains.research.testgenie.llm.error.LLMErrorManager
-import org.jetbrains.research.testgenie.llm.Pipeline
-import org.jetbrains.research.testgenie.llm.SettingsArguments
+import org.jetbrains.research.testgenie.tools.llm.error.LLMErrorManager
+import org.jetbrains.research.testgenie.tools.llm.Pipeline
+import org.jetbrains.research.testgenie.tools.llm.SettingsArguments
 
 class GenerateTestsForClassLLM : AnAction() {
     /**
