--- src/main/kotlin/org/jetbrains/research/testgenie/llm/test/TestLine.kt
@@ -1,7 +1,0 @@
-package org.jetbrains.research.testgenie.llm.test
-
-data class TestLine(
-    val type: TestLineType,
-    val text: String
-) {
-}
