+++ src/main/kotlin/org/jetbrains/research/testgenie/tools/llm/test/TestLine.kt
@@ -1,0 +1,7 @@
+package org.jetbrains.research.testgenie.tools.llm.test
+
+data class TestLine(
+        val type: TestLineType,
+        val text: String
+) {
+}
